requirements.txt
asgi-redis==1.3.0
asgiref==1.1.1
attrs==16.3.0
autobahn==0.18.2
Automat==0.5.0
beautifulsoup4==4.6.0
boto3==1.4.7
botocore==1.7.3
certifi==2017.7.27.1
channels==1.1.3
chardet==3.0.4
constantly==15.1.0
daphne==1.2.0
Django==1.11
django-cors-headers==2.0.2
django-filter==1.0.2
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
docutils==0.14
docxtpl==0.3.2
Flask==0.10.1
idna==2.6
incremental==16.10.1
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.3
lxml==4.0.0
Markdown==2.6.8
MarkupSafe==0.23
msgpack-python==0.4.8
olefile==0.44
Pillow==4.0.0
psycopg2==2.6.2
PyJWT==1.4.2
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.2
redis==2.10.5
requests==2.18.4
s3transfer==0.1.11
six==1.10.0
Twisted==17.1.0
txaio==2.7.0
urllib3==1.22
websocket-client==0.40.0
Werkzeug==0.11.15
zope.interface==4.4.0
